Split image/media_type.py — extract demo data + loader
The 1641-LOC file was 84% demo-dataset code (category constants +
DemoDataset catalog builder + per-source download/embed dispatcher)
sitting on a slim MediaType class.  Pulls the demo code out into two
private helpers so media_type.py focuses on the MediaType contract:

  - _demo_categories.py — Places365 raw text + parser + the six
    Caltech/Food101/EuroSAT/Stanford-Dogs/Oxford-Flowers/UCSF category
    lists (814 LOC of pure data).
  - _demo_sources.py — build_demo_datasets() returning the DemoDataset
    list, plus load_demo_source() that dispatches to the per-source
    downloader + embedder (716 LOC).
  - media_type.py — slim ImageMediaType (173 LOC) that delegates
    demo_datasets / load_demo_source to the helper module.

External users:
  - scripts/run_hac_tree_sweep.py imported _PLACES365_CATEGORIES_LIST
    from media_type; now imports PLACES365_CATEGORIES from
    _demo_categories.
  - The TestPlaces365CategoriesList test referenced
    ImageMediaType._PLACES365_CATEGORIES; now reads the module-level
    PLACES365_CATEGORIES constant.

No behaviour change; pure move.  Closes step 6 of the codebase-reorg
plan (the image/media_type.py audit + split).
